+++ tests/tenancy/test_mt_bulk_create_different_db.py
@@ -51,14 +51,16 @@ class Product(edgy.Model):
 
 registry = copy.copy(models)
 registry.database = another_db
+for val in registry.models.values():
+    val.database = another_db
 
 
 @pytest.fixture(autouse=True, scope="function")
 async def create_test_database():
     async with models:
         await models.create_all()
-        registry.metadata = models.metadata
-        await registry.create_all(False)
+        registry.metadata_by_name = models.metadata_by_name.copy()
+        await registry.create_all(refresh_metadata=False)
         yield
         if not database.drop:
             await models.drop_all()